Fixed turn order in gamemanager
+ Fixed enemy friendly fire
+ Fixed gridspawner intial position tied to its real world location
+ Fixed combat manager damage exploit

+++ Rat_Conquest/Source/Rat_Conquest/Managers/CombatManager/CombatManager.h
@@ -17,7 +17,7 @@ public:
 	// Sets default values for this actor's properties
 	ACombatManager();
 	void DealDamageToUnit(APlayerUnit* Attackerunit, APlayerUnit* Defenderunit);
-	void TakeDamage(APlayerUnit* unit, float amount);
+	void TakeDamage(APlayerUnit* unit, int amount);
 	void KillUnit(APlayerUnit* unit);
 
 protected:
